+++ partiql-lang/src/main/kotlin/org/partiql/lang/planner/transforms/plan/PlanUtils.kt
@@ -1,0 +1,51 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates.  All rights reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License").
+ *  You may not use this file except in compliance with the License.
+ *  A copy of the License is located at:
+ *
+ *       http://aws.amazon.com/apache2.0/
+ *
+ *  or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
+ *  language governing permissions and limitations under the License.
+ */
+
+package org.partiql.lang.planner.transforms.plan
+
+import org.partiql.plan.Attribute
+import org.partiql.plan.Rel
+import org.partiql.plan.Rex
+import org.partiql.types.StaticType
+
+internal object PlanUtils {
+    internal fun getSchema(input: Rel): List<Attribute> = when (input) {
+        is Rel.Project -> input.common.schema
+        is Rel.Aggregate -> input.common.schema
+        is Rel.Bag -> input.common.schema
+        is Rel.Fetch -> input.common.schema
+        is Rel.Filter -> input.common.schema
+        is Rel.Join -> input.common.schema
+        is Rel.Scan -> input.common.schema
+        is Rel.Sort -> input.common.schema
+        is Rel.Unpivot -> input.common.schema
+    }
+
+    internal fun Rex.addType(type: StaticType): Rex = when (this) {
+        is Rex.Agg -> this.copy(type = type)
+        is Rex.Binary -> this.copy(type = type)
+        is Rex.Call -> this.copy(type = type)
+        is Rex.Collection.Array -> this.copy(type = type)
+        is Rex.Collection.Bag -> this.copy(type = type)
+        is Rex.Id -> this.copy(type = type)
+        is Rex.Lit -> this.copy(type = type)
+        is Rex.Path -> this.copy(type = type)
+        is Rex.Query.Collection -> this.copy(type = type)
+        is Rex.Query.Scalar.Pivot -> this.copy(type = type)
+        is Rex.Query.Scalar.Subquery -> this.copy(type = type)
+        is Rex.Switch -> this.copy(type = type)
+        is Rex.Tuple -> this.copy(type = type)
+        is Rex.Unary -> this.copy(type = type)
+    }
+}